requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.2.10
djangorestframework==3.8.2
idna==2.10
mysqlclient==2.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
redis==3.5.3
django-cors-headers==3.5.0
django-redis==4.12.1